refactor(api): extract shared infrastructure into focused modules (#331)
Move auth/CORS/security helpers, Pydantic models, dotenv/SPA helpers,
and lazy-init service singletons out of api_server.py into four focused
modules under src/api/. api_server.py becomes a thin assembler that
creates the FastAPI app, mounts middleware, registers route modules,
and re-exports symbols for test compatibility.

- security.py: auth deps, CORS parsing, DNS-rebinding guard, loopback
  host validation, shell-tools gating (379 lines)
- models.py: Artifact, BacktestMetrics, RAGSelection, RunInfo,
  RunResponse Pydantic schemas (91 lines)
- helpers.py: path constants, dotenv I/O, SPA deep-link middleware,
  path-parameter validation (188 lines)
- state.py: lazy-init session service and channel runtime (124 lines)

api_server.py: 1,103 → 363 lines. Zero behavior change.
All 171 existing API tests pass unchanged.

Closes #331

AI-Model: alibaba-cn/glm-5.2
AI-Contributed/Feature: 1718/1718
AI-Contributed/UT: 57/57
